Refactor factory methods
